More mixin cleanup
