+++ src/main/java/org/orecruncher/dsurround/mixins/audio/MixinSoundEvent.java
@@ -1,4 +1,4 @@
-package org.orecruncher.dsurround.mixins.core;
+package org.orecruncher.dsurround.mixins.audio;
 
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
