+++ src/main/java/org/orecruncher/dsurround/mixins/audio/MixinSoundManagerAccessor.java
@@ -1,4 +1,4 @@
-package org.orecruncher.dsurround.mixins.core;
+package org.orecruncher.dsurround.mixins.audio;
 
 import net.minecraft.client.sound.SoundManager;
 import net.minecraft.client.sound.SoundSystem;
