+++ src/main/java/org/orecruncher/dsurround/mixins/core/MixinRainSplashParticle.java
@@ -16,7 +16,7 @@ import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 public class MixinRainSplashParticle {
 
     @Inject(method = "createParticle(Lnet/minecraft/particle/DefaultParticleType;Lnet/minecraft/client/world/ClientWorld;DDDDDD)Lnet/minecraft/client/particle/Particle;", at = @At("RETURN"), locals = LocalCapture.CAPTURE_FAILSOFT, cancellable = true)
-    public void makeParticle(DefaultParticleType defaultParticleType, ClientWorld clientWorld, double x, double y, double z, double g, double h, double i, CallbackInfoReturnable<Particle> cir) {
+    public void dsurround_makeParticle(DefaultParticleType defaultParticleType, ClientWorld clientWorld, double x, double y, double z, double g, double h, double i, CallbackInfoReturnable<Particle> cir) {
         if (WaterRippleHandler.createRippleParticle(clientWorld, cir.getReturnValue(), new Vec3d(x, y, z))) {
             cir.setReturnValue(null);
         }
