--- src/main/java/org/orecruncher/dsurround/mixins/core/MixinSoundSystem.java
@@ -1,51 +1,0 @@
-package org.orecruncher.dsurround.mixins.core;
-
-import net.minecraft.client.sound.Sound;
-import net.minecraft.client.sound.SoundInstance;
-import net.minecraft.client.sound.SoundSystem;
-import net.minecraft.client.sound.WeightedSoundSet;
-import net.minecraft.sound.SoundCategory;
-import net.minecraft.util.Identifier;
-import org.orecruncher.dsurround.Client;
-import org.orecruncher.dsurround.config.Configuration;
-import org.orecruncher.dsurround.runtime.audio.AudioUtilities;
-import org.orecruncher.dsurround.sound.SoundInstanceHandler;
-import org.orecruncher.dsurround.sound.SoundVolumeEvaluator;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
-
-@Mixin(SoundSystem.class)
-public abstract class MixinSoundSystem {
-
-    @Inject(method = "play(Lnet/minecraft/client/sound/SoundInstance;)V", at = @At("HEAD"), cancellable = true)
-    private void dsurround_play(SoundInstance sound, CallbackInfo ci) {
-        try {
-            if (SoundInstanceHandler.shouldBlockSoundPlay(sound))
-                ci.cancel();
-        } catch (final Exception ignore) {
-        }
-    }
-
-    @Inject(method = "getAdjustedVolume(Lnet/minecraft/client/sound/SoundInstance;)F", at = @At("HEAD"), cancellable = true)
-    private void dsurround_getAdjustedVolume(SoundInstance sound, CallbackInfoReturnable<Float> ci) {
-        try {
-            ci.setReturnValue(SoundVolumeEvaluator.getAdjustedVolume(sound));
-        } catch (final Exception ignore) {
-        }
-    }
-
-    @Inject(method = "play(Lnet/minecraft/client/sound/SoundInstance;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/util/math/Vec3d;<init>(DDD)V"), locals = LocalCapture.CAPTURE_FAILEXCEPTION, cancellable = true)
-    private void dsurround_soundRangeCheck(SoundInstance sound, CallbackInfo ci, WeightedSoundSet weightedSoundSet, Identifier identifier, Sound sound2, float f, float g, SoundCategory soundCategory, float h, float i, SoundInstance.AttenuationType attenuationType, boolean isGlobal) {
-        if (Client.Config.soundSystem.enableSoundPruning) {
-            // If not in range of the listener cancel.
-            if (!SoundInstanceHandler.inRange(AudioUtilities.getSoundListener().getTransform().position(), sound, 4)) {
-                Client.LOGGER.debug(Configuration.Flags.BASIC_SOUND_PLAY, () -> "TOO FAR: " + AudioUtilities.debugString(sound));
-                ci.cancel();
-            }
-        }
-    }
-}
