+++ src/main/resources/dsurround.mixins.json
@@ -10,6 +10,9 @@
     "audio.MixinSoundSystem",
     "audio.MixinSource",
     "audio.MixinStaticSoundAccessor",
+    "audio.MixinSoundEvent",
+    "audio.MixinSoundManagerAccessor",
+    "audio.MixinSoundSystemAccessors",
     "core.MixinAbstractSoundInstance",
     "core.MixinBiome",
     "core.MixinBlockState",
@@ -20,10 +23,6 @@
     "core.MixinParticle",
     "core.MixinParticleManager",
     "core.MixinRaycastContextAccessor",
-    "core.MixinSoundEvent",
-    "core.MixinSoundManagerAccessor",
-    "core.MixinSoundSystem",
-    "core.MixinSoundSystemAccessors",
     "core.MixinWorld",
     "core.MixinRainSplashParticle"
   ],
